% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeLibrary.R
\name{calc.NormFactors}
\alias{calc.NormFactors}
\title{normalizeLibrary}
\usage{
calc.NormFactors(counts = win.counts, input.site = input.site,
  IP.site = IP.site)
}
\arguments{
\item{counts}{counts file generated from previous step}

\item{input.site}{The input sites in counts dataframe.}

\item{IP.site}{The IP sites in counts dataframe.}
}
\description{
normalizeLibrary
}
