% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPefficiency.R
\name{effiTest}
\alias{effiTest}
\title{IPefficiency}
\usage{
effiTest(IP.bam = c(f2, f4, f6), Input.bam = c(f1, f3, f5),
  gtf = NULL, sampleName = NULL)
}
\arguments{
\item{IP.bam}{The ip bam file.}

\item{Input.bam}{The input bam file.}

\item{gtf}{The gtf format gene annotation file}

\item{sampleName}{your sample name, only used for names of print plot and table}
}
\description{
IPefficiency
}
