% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakCalling.R
\name{get.win.counts}
\alias{get.win.counts}
\title{PeakCalling}
\usage{
get.win.counts(IP.bam = c(f2, f4, f6, f8), Input.bam = c(f1, f3, f5,
  f7), gtf = win.gtf, strandSpecific = 0, isPairedEnd = F,
  threads = 4)
}
\arguments{
\item{IP.bam}{The ip bam file.}

\item{Input.bam}{The input bam file.}

\item{gtf}{The gtf format gene annotation file}

\item{strandSpecific}{This option is the same in featureCounts, if your data is strandSpecfic, this value should be 1(stranded) or 2(reversely stranded), if not, it should be 0.}

\item{isPairedEnd}{logical. Depends on your bam file is pair end or single end.}

\item{threads}{Threads used for parallel running}
}
\description{
PeakCalling
}
