% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{motifFind}
\alias{motifFind}
\title{segment}
\usage{
motifFind(peakFile = NULL, species = NULL, plotmethod = "prob",
  sortCut = 5000)
}
\arguments{
\item{peakFile}{The m6A peak file}

\item{species}{The species, eg: "mm10", "hg38"}

\item{plotmethod}{The method to plot motif, "prob" or "bit"}

\item{sortCut}{Choose the most significant peaks to find motifs, this value means the amount of peaks you wanna choose}
}
\description{
segmentation algorithm and then map transcript loci from genome loci
}
