% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{reportResults}
\alias{reportResults}
\title{segment}
\usage{
reportResults(gtf = win.gtf, catMethyl = NULL, threads = 4)
}
\arguments{
\item{gtf}{The splited gtf file}

\item{catMethyl}{The output file of sig.test()}

\item{threads}{Threads used for parallel running}
}
\description{
segmentation algorithm and then map transcript loci from genome loci
}
