% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmTest.R
\name{sig.test}
\alias{sig.test}
\title{glmTest}
\usage{
sig.test(counts, condition = c("KO", "WT"), cutoff = 100, rep = 1)
}
\arguments{
\item{counts}{counts file generated from previous step}

\item{condition}{The input bam file.}

\item{cutoff}{The gtf format gene annotation file}

\item{rep}{This option is the same in featureCounts, if your data is strandSpecfic, this value should be 1(stranded) or 2(reversely stranded), if not, it should be 0.}
}
\description{
significant test
}
