% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBin.R
\name{sub.win.gtf}
\alias{sub.win.gtf}
\title{splitBin}
\usage{
sub.win.gtf(gtf, binSize = 100, step = 10, threads = 4)
}
\arguments{
\item{gtf}{The gtf format gene annotation file}

\item{binSize}{Txdb file. Could be NULL when gtf is avaliable}

\item{step}{Step of Sliding window}

\item{threads}{Threads used for parallel running}
}
\description{
split transcriptome to sliding windows with step
}
